refactor: survey 타입 단일화
