+++ app/survey/utils/buildSurveyPayload.ts
@@ -1,35 +1,8 @@
 import type { CategoryKey } from '@/app/survey/data/buildQuestion';
-
-export type IntensityLabel = 'very_low' | 'low' | 'medium' | 'high' | 'very_high';
-
-//전체 payload 구조
-export type SurveyResultPayload = {
-  //기록 정렬
-  createdAt: string;
-
-  //기본 질문
-  basicInfo: {
-    gender: string;
-    ageGroup: string;
-  };
-  //카테고리
-  selectedCategories: CategoryKey[];
-
-  //카테고리별 설문 결과
-  categories: Record<
-    CategoryKey,
-    {
-      //multiChoice 질문 선택 결과
-      state: string[];
-
-      // scaleChoice 결과
-      intensity: IntensityLabel;
-    }
-  >;
-};
+import { Intensity, SurveyResultPayload } from '@/types/survey';
 
 // scaleChoice 숫자 값을 의미값으로 매핑
-const INTENSITY_MAP: Record<number, IntensityLabel> = {
+const INTENSITY_MAP: Record<number, Intensity> = {
   1: 'very_low',
   2: 'low',
   3: 'medium',
@@ -37,7 +10,7 @@ const INTENSITY_MAP: Record<number, IntensityLabel> = {
   5: 'very_high',
 };
 
-function mapIntensity(value: unknown): IntensityLabel {
+function mapIntensity(value: unknown): Intensity {
   const n = Number(value);
   return INTENSITY_MAP[n] ?? 'medium';
 }
